requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django>=3.0,<4.0
django-ckeditor==6.4.2
django-js-asset==2.0.0
Pillow==9.2.0
gunicorn
python-decouple